mipt_overflow/requirements.txt
backports.shutil-get-terminal-size==1.0.0
click==6.7
decorator==4.2.1
Django==1.11.11
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
enum34==1.1.6
ipython==5.5.0
ipython-genutils==0.2.0
mysqlclient==1.3.12
pathlib2==2.3.0
pexpect==4.4.0
pickleshare==0.7.4
pip-chill==0.1.7
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2018.3
scandir==1.7
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
traitlets==4.3.2
wcwidth==0.1.7
